% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{d_laplace}
\alias{d_laplace}
\title{Laplace denstiy function}
\usage{
d_laplace(x)
}
\arguments{
\item{x}{variable(vector)}
}
\value{
the density at \code{x}
}
\description{
Laplace denstiy function using Rcpp
}
\examples{
\dontrun{
x <- seq(1, 3, length.out=10)
pdf <- d_lapalce(); pdf
}
}
