% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatCompR.R
\name{d_mvnorm}
\alias{d_mvnorm}
\title{Multivariate normal distribution density function}
\usage{
d_mvnorm(x, mu, Sigma)
}
\arguments{
\item{x}{at which the density is wanted}

\item{mu}{the mean of the multivariate normal distribution}

\item{Sigma}{the variance of the multivariate normal distribution}
}
\value{
the density at \code{x}
}
\description{
multivariate normal distribution probability density function
}
\examples{
\dontrun{
d_mvnorm(c(0,0), c(0, 0), Sigma=matrix(c(1,0,0,1), ncol=2))
}
}
