% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatCompR.R
\name{Euclidean.dist}
\alias{Euclidean.dist}
\title{knn use kernel density function}
\usage{
Euclidean.dist(x, y)
}
\arguments{
\item{x}{data 1}

\item{y}{data 2}
}
\value{
distance between \code{x} and \code{y}
}
\description{
knn use kernel density function
}
\examples{
\dontrun{
x <- rnorm(12)
y <- rnorm(12, 1, 3)
Eucildean.dist(x, y)
}
}
