% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatCompR.R
\name{Euclidean.norm}
\alias{Euclidean.norm}
\title{A Gibbs sampler using R}
\usage{
Euclidean.norm(x)
}
\arguments{
\item{x}{a vector}
}
\value{
the L2 norm of \code{x}
}
\description{
A Gibbs sampler using R
}
\examples{
\dontrun{
x <- rnorm(12)
d <- Euclidean.norm(x)
d
}
}
