% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatCompR.R
\name{lck.reg}
\alias{lck.reg}
\title{Local constant kernel regression}
\usage{
lck.reg(x, X, Y, h)
}
\arguments{
\item{x}{at which the density is estimated}

\item{X}{sample}

\item{Y}{sample}

\item{h}{band width}
}
\value{
the density at \code{x}
}
\description{
local constant kernel regression
}
\examples{
\dontrun{
X <- runif(200, 0, 15)
Y <- (X-1.4)*(X-4.6)*(X-11.2) + rnorm(200, 0, 20)
x <- runif(10, 0, 10)
lck.reg(x, X, Y, 0.2)
}
}
