% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatCompR.R
\name{r_mvnorm}
\alias{r_mvnorm}
\title{Multivariate normal distribution random number}
\usage{
r_mvnorm(n, mu, Sigma)
}
\arguments{
\item{n}{the size of random number we need}

\item{mu}{the mean of the multivariate normal distribution}

\item{Sigma}{the variance of the multivariate normal distribution}
}
\value{
the random numbers yield N(mu, Sigma)
}
\description{
generate multivariate normal distribution random number
}
\examples{
\dontrun{
x <- r_mvnorm(100, mu=c(1, 2), Sigma=matrix(c(1,0,0,2), ncol=2))
mean.hat <- apply(x, 2, mean)
var.hat <- apply(x, 2, var)
}
}
