% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rw_Metropolis}
\alias{rw_Metropolis}
\title{Metropolis method}
\usage{
rw_Metropolis(sigma, x0, N)
}
\arguments{
\item{sigma}{vaiance}

\item{x0}{initial values}

\item{N}{length}
}
\value{
a list of \code{x} and iterator \code{k}
}
\description{
using Metropolis method implementing Mcmc
}
\examples{
\dontrun{
sigma = 6.
x0 = 36.
N = 2000
model <- rw_Metropolis(sigma, x0, N)$x
plot(model)
}
}
