\name{StatComp20027-package}
\alias{StatComp20027-package}
\docType{package}
\title{
  This package provides several functions used for my Statistical Computing and Nonparametric courses.
}
\description{
  \code{Euclidean.dist} and \code{Euclidean.norm} functions are provided in this package.
  The package created mainly as a final project of the __Statistical Computing__ course.
}
\author{
WANG Yu

Maintainer: WANG Yu <theyu@mail.ustc.edu.cn>
}
\keyword{ package }
